Fixed FillScreen.
Fixed fillscreen to work on matrix_manager. This was using variables that I never ended up using, causing it to fail.

One more bug exists: atexit doesn't work inside objects, so will have to move the exit function to outside of the class, or make matrix a persistent object instead of an instance.