Implemented static tile display.
Static tiles are tiles that have no scrolling and therefore cannot have an automatic max frames calculation.

TileManager calculates what the max frames should be using a user-defined lifespan for static tiles and the estimated FPS value that they want the display to run at. Tiles are then set a max frame value which corresponds to approximately the time the user wanted static tiles to display.

Adjusted _AllTilesDisplayed() logic to ensure that even if the displayed bit is set, all frames have been displayed (would have been a bug for the last tile to display, it would have displayed 1 frame and then exited the run loop).